增加一些验证,数据库方面基本确定,还有些需要再考虑

+++ src/main/java/com/nwuer/action/AcademyAction.java
@@ -2,6 +2,8 @@ package com.nwuer.action;
 
 import java.util.List;
 
+import javax.servlet.http.HttpServletRequest;
+
 import org.apache.struts2.ServletActionContext;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Scope;
@@ -9,6 +11,7 @@ import org.springframework.stereotype.Controller;
 
 import com.nwuer.entity.Academy;
 import com.nwuer.service.AcademyService;
+import com.nwuer.utils.ValidateUtil;
 import com.opensymphony.xwork2.ActionSupport;
 import com.opensymphony.xwork2.ModelDriven;
 
@@ -21,8 +24,12 @@ public class AcademyAction extends ActionSupport implements ModelDriven<Academy>
 		return academy;
 	} //模型驱动获取数据
 	
+	String info;
+	
 	@Autowired
 	private AcademyService academyService;
+	@Autowired
+	private ValidateUtil validateUtil; 
 	
 	/**
 	 * 展示添加院系
@@ -40,19 +47,34 @@ public class AcademyAction extends ActionSupport implements ModelDriven<Academy>
 	 * @return
 	 */
 	public String add() {
-		//验证
+		//验证   a_number
+		info = validateUtil.validateNumber(this.academy.getA_number(),10 );
+		HttpServletRequest req = ServletActionContext.getRequest();
+		if(info != null) {
+			req.setAttribute("info", "编号"+info);
+			return ERROR;
+		}
+			
+		
 		try {
 			this.academyService.add(academy);
 		}catch(Exception e) {
-			String info = "此院系下还有其他专业,请先清理此院系下的专业";
-			
+			info = "此院系下还有其他专业,请先清理此院系下的专业";
+			req.setAttribute("info", info);
+			return ERROR;
 		}
 		
 		return SUCCESS;
 	}
 	
 	public String delete() {
-		//验证
+		//从数据库验证是否存在验证 
+		Academy a = this.academyService.getByIdEager(this.academy.getA_id());
+		if(a == null) {
+			info = "系统错误,请刷新重试或联系维护人员";
+			ServletActionContext.getRequest().setAttribute("info", info);
+			return ERROR;
+		}
 		
 		this.academyService.delete(this.academy.getA_id());
 		return SUCCESS;
@@ -73,6 +95,12 @@ public class AcademyAction extends ActionSupport implements ModelDriven<Academy>
 	 */
 	public String update() {
 		//验证
+		info = validateUtil.validateNumber(this.academy.getA_number(),10 );
+		HttpServletRequest req = ServletActionContext.getRequest();
+		if(info != null) {
+			req.setAttribute("info", "编号"+info);
+			return ERROR;
+		}
 		
 		Academy a = this.academyService.getById(this.academy.getA_id());
 		a.setA_name(academy.getA_name());
