+++ src/main/java/com/nwuer/action/SubjectAction.java
@@ -13,6 +13,7 @@ import com.nwuer.entity.Major;
 import com.nwuer.entity.Subject;
 import com.nwuer.service.MajorService;
 import com.nwuer.service.SubjectService;
+import com.nwuer.utils.ValidateUtil;
 import com.opensymphony.xwork2.ActionSupport;
 import com.opensymphony.xwork2.ModelDriven;
 @Controller
@@ -23,15 +24,22 @@ public class SubjectAction extends ActionSupport implements ModelDriven<Subject>
 	public Subject getModel() {
 		return subject;
 	}//模型驱动获取数据
+	String info;
 	
 	@Autowired
 	private SubjectService subjectService;
 	@Autowired
 	private MajorService majorService;
+	@Autowired
+	private ValidateUtil validateUtil;
 	
 	public String delete() {
 		//验证
-		
+		Subject s = this.subjectService.getByIdEager(subject.getSub_id());
+		if(s == null) {
+			ServletActionContext.getRequest().setAttribute("info", "系统错误,请稍后重试");
+			return ERROR;
+		}
 		this.subjectService.delete(this.subject.getSub_id());
 		return SUCCESS;
 	}
@@ -51,6 +59,12 @@ public class SubjectAction extends ActionSupport implements ModelDriven<Subject>
 	
 	public String add() {
 		//验证
+		HttpServletRequest req = ServletActionContext.getRequest();
+		info = validateUtil.validateNumber(subject.getSub_number(), 6);
+		if(info != null) {
+			req.setAttribute("info", info);
+			return ERROR;
+		}
 		
 		int id = this.subjectService.add(subject);
 		if(id > 0) {
@@ -58,12 +72,17 @@ public class SubjectAction extends ActionSupport implements ModelDriven<Subject>
 			return SUCCESS;
 		} else {
 			//添加失败
+			req.setAttribute("info", "系统错误,请稍后重试");
 			return ERROR;
 		}
 	}
 	
 	public String edit() {
 		Subject s = this.subjectService.getById(this.subject.getSub_id());
+		if(s == null) {
+			ServletActionContext.getRequest().setAttribute("info", "系统错误,请稍后重试");
+			return ERROR;
+		}
 		List<Major> list = this.majorService.getAll();
 		
 		HttpServletRequest req = ServletActionContext.getRequest();
@@ -74,6 +93,10 @@ public class SubjectAction extends ActionSupport implements ModelDriven<Subject>
 	
 	public String update() {
 		Subject s = this.subjectService.getById(subject.getSub_id());
+		if(s == null) {
+			ServletActionContext.getRequest().setAttribute("info", "系统错误,请稍后重试");
+			return ERROR;
+		}
 		subject.setCreate_time(s.getCreate_time());
 		
 		this.subjectService.update(subject);
